refactor confirmationBox
