+++ ClientApp/src/components/admin/dashboard-all/RecipientTable.tsx
@@ -31,7 +31,7 @@ import DeleteTypeDialog from "components/admin/dashboard-all/DeleteTypeDialog";
 import EditFamilyDialog from "components/shared/EditFamilyDialog";
 import SendEmailContent from "components/shared/SendEmailContent";
 import SendIcon from "@material-ui/icons/Send";
-import ConfirmationBox from "components/shared/confirmationBox";
+import ConfirmationBox from "components/shared/ConfirmationBox";
 import ApiService from "common/functions/apiServiceClass";
 
 type Props = {
